LayoutTests/platform/wpe/fast/text/font-stretch-parse-expected.txt

PASS window.getComputedStyle(document.getElementById('test1')).fontStretch is "normal"
PASS window.getComputedStyle(document.getElementById('test2')).fontStretch is "1%"
PASS window.getComputedStyle(document.getElementById('test3')).fontStretch is "normal"
PASS window.getComputedStyle(document.getElementById('test4')).fontStretch is "ultra-condensed"
PASS window.getComputedStyle(document.getElementById('test5')).fontStretch is "extra-condensed"
PASS window.getComputedStyle(document.getElementById('test6')).fontStretch is "condensed"
PASS window.getComputedStyle(document.getElementById('test7')).fontStretch is "semi-condensed"
PASS window.getComputedStyle(document.getElementById('test8')).fontStretch is "normal"
PASS window.getComputedStyle(document.getElementById('test9')).fontStretch is "semi-expanded"
PASS window.getComputedStyle(document.getElementById('test10')).fontStretch is "expanded"
PASS window.getComputedStyle(document.getElementById('test11')).fontStretch is "extra-expanded"
PASS window.getComputedStyle(document.getElementById('test12')).fontStretch is "ultra-expanded"
PASS window.getComputedStyle(document.getElementById('test13')).fontStretch is "normal"
PASS window.getComputedStyle(document.getElementById('test14')).fontStretch is "normal"
PASS window.getComputedStyle(document.getElementById('test15')).fontStretch is "normal"
PASS window.getComputedStyle(document.getElementById('test16')).fontStretch is "extra-condensed"
PASS window.getComputedStyle(document.getElementById('test17')).fontStretch is "extra-condensed"
PASS window.getComputedStyle(document.getElementById('test18')).fontStretch is "normal"
PASS window.getComputedStyle(document.getElementById('test19')).fontStretch is "normal"
PASS window.getComputedStyle(document.getElementById('test20')).fontStretch is "extra-expanded"
PASS window.getComputedStyle(document.getElementById('test21')).fontStretch is "7%"
PASS window.getComputedStyle(document.getElementById('test1')).font is "normal normal normal normal 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test2')).font is "16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test3')).font is "normal normal normal normal 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test4')).font is "normal normal normal ultra-condensed 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test5')).font is "normal normal normal extra-condensed 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test6')).font is "normal normal normal condensed 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test7')).font is "normal normal normal semi-condensed 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test8')).font is "normal normal normal normal 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test9')).font is "normal normal normal semi-expanded 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test10')).font is "normal normal normal expanded 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test11')).font is "normal normal normal extra-expanded 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test12')).font is "normal normal normal ultra-expanded 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test13')).font is "normal normal normal normal 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test14')).font is "normal normal normal normal 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test15')).font is "normal normal normal normal 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test16')).font is "normal normal 100 extra-condensed 48px/49px \"Helvetica Neue\""
PASS window.getComputedStyle(document.getElementById('test17')).font is "normal normal 100 extra-condensed 48px/49px \"Helvetica Neue\""
PASS window.getComputedStyle(document.getElementById('test18')).font is "normal normal 100 normal 48px/49px \"Helvetica Neue\""
PASS window.getComputedStyle(document.getElementById('test19')).font is "normal normal 100 normal 48px/49px \"Helvetica Neue\""
PASS window.getComputedStyle(document.getElementById('test20')).font is "italic small-caps 100 extra-expanded 48px/49px \"Helvetica Neue\""
PASS window.getComputedStyle(document.getElementById('test21')).font is "16px/18px Times"
PASS document.getElementById('test1').style.font is ""
PASS document.getElementById('test16').style.font is "100 extra-condensed 48px/49px \"Helvetica Neue\""
PASS document.getElementById('test17').style.font is "100 extra-condensed 48px/49px \"Helvetica Neue\""
PASS document.getElementById('test18').style.font is "100 48px/49px \"Helvetica Neue\""
PASS document.getElementById('test19').style.font is "100 48px/49px \"Helvetica Neue\""
PASS document.getElementById('test20').style.font is "italic small-caps 100 extra-expanded 48px/49px \"Helvetica Neue\""
PASS successfullyParsed is true

TEST COMPLETE
